# Native division for UInt128 on 32bit platforms