I'm sure there are clever tricks to pull off, but it seems to be an improvement against master already, and this not quite my field of expertise. Note that these benchmarks are made with a 32bit-julia (MARCH=i686) on a 64bit system.

Master:
```julia
julia> @btime $(one(UInt128)) ÷ $(one(UInt128));
  757.851 ns (11 allocations: 156 bytes)

julia> @btime $(-one(UInt128)) ÷ $(one(UInt128));
  868.255 ns (12 allocations: 180 bytes)

julia> @btime $(one(UInt128)) % $(one(UInt128));
  296.552 ns (8 allocations: 60 bytes)

julia> @btime $(-one(UInt128)) % $(one(UInt128));
  318.302 ns (8 allocations: 72 bytes)

julia> @eval Base function div(x::UInt128, y::UInt128)
           return UInt128(div(BigInt(x), BigInt(y)))::UInt128
       end;
       
julia> one(UInt128) ÷ one(UInt128);

julia> @btime $(one(UInt128)) ÷ $(one(UInt128));
  344.019 ns (10 allocations: 124 bytes)

julia> @btime $(-one(UInt128)) ÷ $(one(UInt128));
  451.320 ns (11 allocations: 148 bytes)

julia> @btime $(one(UInt128)) % $(one(UInt128));
  332.386 ns (10 allocations: 124 bytes)

julia> @btime $(-one(UInt128)) % $(one(UInt128));
  355.200 ns (10 allocations: 136 bytes)
```
The recursion limiting heuristics make it so that the results depend on which function is compiled first, so re-evaluating the original definition yields different results.

This PR:
```julia
julia> @btime $(one(UInt128)) ÷ $(one(UInt128));
  0.628 ns (0 allocations: 0 bytes)

julia> @btime $(-one(UInt128)) ÷ $(one(UInt128));
  383.621 ns (0 allocations: 0 bytes)

julia> @btime $(one(UInt128)) % $(one(UInt128));
  0.628 ns (0 allocations: 0 bytes)

julia> @btime $(-one(UInt128)) % $(one(UInt128));
  301.164 ns (0 allocations: 0 bytes)
```

I'm not 100% sure `0xfff...fff % 0x000...001` shows the worst-case performance, so there might be inputs where current master is slightly faster, but in any case, we get rid of the allocations.