@martinholters 2018-11-29
The slight simplification of `rem` gives a 10% improvement here:
```julia
julia> @btime $(-one(UInt128)) % $(one(UInt128));
  277.050 ns (0 allocations: 0 bytes)
```