@martinholters 2018-12-12
@simonbyrne, any thoughts on this? Anyone else willing to look at this? (I've tried a few minor modifications, none of which seemed to be a clear improvement.)