@simonbyrne 2018-12-12
Sorry, this is probably a bit too far out of my expertise for me to review beyond "this would be good to have".