@StefanKarpinski 2018-12-12
Do we have enough tests of 128-bit `divrem` to be confident that this is correct? If so, I think we can go ahead and merge this. Your call, @martinholters.