@stevengj 2018-12-13
Do you have any reference for the algorithm you are using that you can add as a comment to the code?

The code is too hairy for me to easily proofread, so my main concern is that we have test coverage of all of the branches here.