@staticfloat 2018-12-13
From a slightly different point of view, I like this because it decreases our dependence on GMP/MPFR, which helps for restrictive environments that don't allow those packages due to licensing reasons.