@staticfloat 2018-12-14
Well, as far as correctness goes, it looks pretty good to me:

```
# <paste function here as `divrem2()`>

@testset "Random integers" begin
    for idx in 1:100000000
        x, y = rand(UInt128, 2)
        @test divrem2(x, y) == Base.divrem(x,y)
    end
end
```

That test took about 5 minutes to run, and it passed all cases, so that makes me pretty happy.