@martinholters 2018-12-14
Yes, I have made similar randomized tests and they look good. However, getting good test coverage of the possible branches is pretty tough. E.g. I uncovered some bugs only by explicitly using small (random) divisors.

I have an idea for an algorithm which might be a bit easier to grok and first tests indicate that performance might be at least as good, but I'm not sure I'll find time to pursue this before January. But then again, there is no hurry here.