@stevengj 2018-12-17
You'd need to explicitly design tests to cover the corner cases of small divisors etc.