@martinholters 2019-01-08
>I have an idea for an algorithm which might be a bit easier to grok and first tests indicate that performance might be at least as good

Nope, turned out to be slower. Instead, I've tried to make the existing algorithm more readable. That called for introducing a function for setting the `i`-th bit in a `UInt128`. (`x |= UInt128(1) << i` leads to vastly sub-optimal code for the 32bit target.) I've named it `_setbit` for now, limited it to `UInt128`, and it's only defined on 32bit targets. Would it make sense to give it an underscore-free name and (unconditionally) provide the fallback `x | one(typeof(x)) << i` definition?

Also, I've added some tests.